allow more gpu retries
This is done for SignalP. When there is insufficient gpu vmem for signalp it auto-falls back to the cpu. However, our data directories only contain gpu Signalp models so signalp then throws an  error saying it can't run gpu models on a cpu and it crashes, in turn causing iprscan6 to crash. So we need to tolerate this happening a few times, especially when te gpus are busy